Added Week01 exercise:
Adjust the game loop in your project to your liking based on what was discussed.
Introduce a GameObject-Component system in your project
Make sure to provide methods so a user can
Add a component to a game object
Remove a component from a game object in a safe manner
Get a component from a game object
Check whether a component has been added
Think about this: should it ever be possible for a component to exist without a gameobject?
Run a correct Update method loop in your project.
Add an FPS counter